chore: reorg in filter configuration
